+++ src/file_reader.h
@@ -1,8 +1,8 @@
 #ifndef FILE_READER_H
 #define FILE_READER_H
 
-#include <stdint.h>
+#include <stddef.h>
 
-unsigned char *read_file(const char *file_name, uint64_t *file_size);
+unsigned char *read_file(const char *file_name, size_t *file_size);
 
 #endif // FILE_READER_H
